refactor: reorganize components and update imports for Navbar and ThemeProvider
